requirements.txt
appdirs==1.4.3
asgiref==3.2.10
autopep8==1.5.4
CacheControl==0.12.6
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.3
contextlib2==0.6.0
cryptography==3.4.7
defusedxml==0.7.1
distlib==0.3.0
distro==1.4.0
dj-database-url==0.5.0
dj-rest-auth==2.1.4
dj-static==0.0.6
Django==3.1.1
django-allauth==0.44.0
django-cors-headers==3.7.0
django-taggit==1.3.0
django-taggit-serializer==0.1.7
djangorestframework==3.11.1
gunicorn==20.0.4
html5lib==1.0.1
idna==2.10
ipaddr==2.2.0
lockfile==0.12.2
model-mommy==2.0.0
msgpack==0.6.2
oauthlib==3.1.0
packaging==20.3
pep517==0.8.2
Pillow==7.2.0
progress==1.5
pycodestyle==2.6.0
pycparser==2.20
PyJWT==2.0.1
pyparsing==2.4.6
python-decouple==3.4
python3-openid==3.2.0
pytoml==0.1.21
pytz==2020.1
requests==2.25.1
requests-oauthlib==1.3.0
retrying==1.3.3
six==1.15.0
sqlparse==0.3.1
static3==0.7.0
toml==0.10.1
urllib3==1.26.4
webencodings==0.5.1
whitenoise==5.2.0
